* 微小修改
git-svn-id: svn://zwd33.gicp.net/DVM3/trunk@12 57787194-4ec7-4324-af3d-b87abe402d06

+++ Sources/DvmTests/TestBase.cs
@@ -23,14 +23,14 @@ namespace DvmTests
 
 		#region Sleep
 
-		protected static void Sleep(float seconds)
+		protected static void Sleep(double seconds)
 		{
 			Thread.Sleep((int)(seconds * 1000));
 		}
 
 		protected static void Sleep()
 		{
-			Sleep(0.5f);
+			Sleep(0.5);
 		}
 
 		#endregion
